libs/pdf-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-pipeline"
version = "0.1.0"
description = "Generic PDF-to-JSON extraction pipeline via LLM"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "google-cloud-vision>=3.5.0",
    "pypdfium2>=4.0.0",
    "Pillow>=10.0.0",
    "xlwings>=0.30.0",
    "openpyxl>=3.0.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools.packages.find]
where = ["src"]
